cleanup extra tests and functions for transactions
